Add feature flag to block starting VMs
The flag is added to the `enabled_when_unknown` list, which tells xapi
to enable the feature if the flag is absent from `host.license_params`
(as opposed to present and `true` or `false`). This is done to ensure
that VM start is not inadvertently blocked during the rolling pool
update that introduces this change.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>