CP-50078: Determine if SMAPI observer is enabled
As part of adding tracing to xapi-storage-script, we determine whether
observation is enabled for the SMAPI component by checking for the presence of
*.observer.conf files in the configuration directory. In usual circumstances,
these checks - when initiated from within xapi itself - would query the
database. However, as xapi-storage-script is an external daemon, we use
external means. Future changes may wish to endow the xapi-storage-script
interface (within xapi-idl) to be aware of tracing and enable and disable
tracing that way.

To avoid having to propagate all of this to the intended place (fork_exec_rpc)
via parameters, a minimal configuration record is created. This may change in
future but it is a convenient bit of global state for keeping the changeset
small.

Signed-off-by: Colin James <colin.barr@cloud.com>