CP-50053: Functionalise external authentication
In preparation for memoization, we wrap a region of
"login_with_password" (the part calling out to external authentication code)
within a nested function ("query_external_auth").

We introduce a new record type "external_auth_result" that represents the set
of downward exposed variables that that are bound within the code our nested
function has incorporated. This record becomes the result of
"query_external_auth". Consequently, the function is immediately invoked and
its result is matched upon to bring all of the fields into scope. The future
memoization code will treat a call to this function as the "slow path".

The extent of the wrapping function is rather large. Until we have tracing on
the related external authentication functions, we conservatively seek to
memoize a large region of login_with_password.

Signed-off-by: Colin James <colin.barr@cloud.com>