CP-50053: Add authentication cache
Provides a simple authentication cache built using a bounded priority search
queue.

The structure of the cache is effectively a (bounded) priority search queue,
mapping usernames to entries - where entries have an explicit time-to-live and
can store arbitrary data (optionally secured with a cryptographic hash).

Also provides a configurable expiry span in the form of a global in Xapi_globs,
"external_authentication_expiry_ms". This time span is added to the current
time - upon entering data into the cache - to provide an expiration time. Its
default value is 300,000 (milliseconds), i.e. 5 minutes.

The underlying priority search queue treats the minimum entry as being the
entry with highest priority. Therefore, as entries are compared by their
expiration time, the entry that is soonest to expire is the one that is evicted
when attempting to add a new entry when the cache is at capacity.

Signed-off-by: Colin James <colin.barr@cloud.com>