CP-50053: Add authentication cache options
Adds configurable options "enable-external-authentication-cache" and
"external-authentication-cache-size" to Xapi_globs, which control whether the
cache is enabled and the maximum capacity of the cache, respectively.

Also adds "external-authentication-expiry-ms" option, which permits
configuration of the time-to-live assigned to each cache entry. The global that
it manipulates in Xapi_globs was introduced earlier.

Signed-off-by: Colin James <colin.barr@cloud.com>