CP-50053: Memoize external authentication
Provides a first-class modules approach for deciding between cache
implementations at runtime. The cache is initialised upon first use and stored
in a modular container, alongside its implementation module.

Rewrites the call to "query_external_auth" to use a memoizing wrapper that only
calls "query_external_auth" in the slow path (when no valid cache entry
exists).

It may be decided that the choice of cache need not be as flexible, thus
negating any benefit of using first-class modules. I've provided it anyway as a
nice design pattern for switching between modular implementations at runtime
(with a common interface, in this case: EXTERNAL_AUTH_CACHE). Alternative
approaches can: (1) just manipulate a X.t option, where X is the hardcoded
choice of module (with relevant casing to handle when a cache is and isn't
present). Or, (2) manipulate X.t where X is a hardcoded choice of module, with
the invariant that a concrete cache implementation with capacity ("size") of 0
is logically equivalent to having no cache.

Signed-off-by: Colin James <colin.barr@cloud.com>