CP-50422: Destroy authentication cache in disable_external_auth
Upon successfully disabling external authentication (e.g. winbind), any extant
external authentication cache is cleared (as to not undermine disabling of the
external authentication route).

We consider invalidating the cache as sufficient for our needs because external
authentication would need to succeed for the cache to become populated again
(even if it is recreated due to being enabled as a feature generally).

Signed-off-by: Colin James <colin.barr@cloud.com>