CP-51034: Add cache-related fields to pool object
Adds 3 new fields to the pool object:

- ext_auth_cach_enabled: bool

  Specifies whether or not external authentication caching is enabled for this
  pool. The field is false by default.

- ext_auth_cach_size: int

  Specifies the maximum capacity of the external authentication cache. Its
  default value is 50.

- ext_auth_cache_expiry: int

  Specifies how long cached entries should remain cached. The default value is
  300 seconds (5 minutes).

Previously, these options were configured from xapi.conf, but in the case of
designation of new pool coordinator (by autonomous means or otherwise), we
would like for the feature's configuration to be adopted by new coordinator
(without necessitating duplicate xapi configurations across hosts in the pool).
The cache's contents itself will not be preserved, but caching will resume on
the new coordinator if it was enabled on the previous coordinator.

Signed-off-by: Colin James <colin.barr@cloud.com>