IH-631: Only case on alternative wire name
To avoid having a duplicated numbers of logical cases in server.ml, we
favour dispatching over the alternative wire name only.

For example, previously a case would look like:

| "class.foo_bar" | "class_foo_bar" ->

Where the latter case in this disjunctive pattern is the "alternative"
wire format (using an underscore instead of a period).

Now, we normalise the __call name by way of replacing the first
occurrence of "." with "_" and dispatch over that. Hardcoded cases such
as system.listMethods and system.isAlive are rewritten to match this
format.

E.g.

let __normalised = Str.(replace_first (regexp "\\.") "_" __call) in
...
begin match __normalised with
  ...
  | "class_foo_bar" -> ...
  ...
end

This is quite an opinionated change as string pattern matching is
special-cased within the OCaml compiler (see strmatch.ml). Halving the
number of logical cases does not necessarily result in a halving of the
amount of branching code that is emitted by the compiler. In brief, the
OCaml runtime represents strings as an array of words (with the last
word containing padding zeroes). In matching, the cases are first
discriminated by length, and then discriminated via binary searching
over the values of their components (words) at different indices. The
order in which the words are discriminated is up to a variance
heuristic. Only in the worst case - where there are no overlapping cases
at each word position - would halving the logical number of cases result
in halving the amount of branching code generated. In our case, there is
a lot of word overlap (arising from common prefixes such as "Class.get",
etc.) that is effectively matched in parallel.

It is simpler to experiment with alternative dispatch mechanisms - such
as hash tables - if each message name corresponds to only one dispatch
name.

Signed-off-by: Colin James <colin.barr@cloud.com>