IH-631: Add xapi-dispatchers library machinery
Introduces some of the required dynamic_include machinery for
dynamically generating sub-dispatchers for use by server.ml.

The idea is that each generated sub-dispatcher will be responsible for
dispatching over the messages of a subset of the classes in xapi. Then.
the top-level dispatcher (server.ml) will case on the class names in
order to decide which generated sub-dispatcher module (server_i.ml) to
call.

The general approach is to define a subdirectory xapi-generated which
will contain a generated dune file (comprising the rules to build
sub-dispatchers) that is then dynamically included by a new library,
called xapi-dispatchers, will will expose itself as unwrapped and only
be linked by the internal "server only" library.

Signed-off-by: Colin James <colin.barr@cloud.com>