Restructure style of From.add
The current style uses a lot of nesting (and a shadow of `s`) and, due
to the formatting rules, the code is awkwardly indented to the right.
Instead, filter for events (performing the relevant side effect of
updatnig the event number) without a fold.

The resulting sequence being non-empty implies a match, rather than
propagating `true` through a fold.

Signed-off-by: Colin James <colin.barr@cloud.com>