Use records when accumulating events
Rather than accumulating the components of a tuple, we accumulate the
fields of a record. The benefit is that code is made clearer by
functional update syntax, where we explicitly cite the modified fields.

Futher, smaller, structural changes are also introduced to the related
code. In particular, we make last_generation invariant within the
folding procedure. The lack of pervasive mutability simplifies things a
bit.

The folding procedure is composed of two explicit helpers which fold
recent database entries (to affect create and modify events) and deleted
database entries (to affect delete events). These are then used in
reverse application style to make the accumulative effect clearer.

Signed-off-by: Colin James <colin.barr@cloud.com>